Real Time Fraud Detection/requirements.txt
kafka-python==2.0.2
streamlit==1.31.0
pandas==2.2.0
altair==5.2.0